commands: move backout code to backout.rs
